+++ crates/wdl-gauntlet/src/lib.rs
@@ -4,7 +4,6 @@
 #![warn(missing_docs)]
 #![warn(rust_2018_idioms)]
 #![warn(rust_2021_compatibility)]
-#![warn(missing_debug_implementations)]
 #![warn(clippy::missing_docs_in_private_items)]
 #![warn(rustdoc::broken_intra_doc_links)]
 
@@ -31,8 +30,8 @@ pub use repository::Repository;
 
 use crate::config::ReportableConcern;
 use crate::report::Status;
-use crate::repository::options;
 use crate::repository::Identifier;
+use crate::repository::WorkDir;
 
 /// The exit code to emit when any test unexpectedly fails.
 const EXIT_CODE_FAILED: i32 = 1;
@@ -55,12 +54,6 @@ pub enum Error {
     /// A WDL 1.x parse tree error.
     GrammarV1(grammar::v1::Error),
 
-    /// An error related to a [`Repository`].
-    Repository(repository::Error),
-
-    /// An error related to a repository [`Builder`](repository::Builder).
-    RepositoryBuilder(repository::builder::Error),
-
     /// An error related to a repository [`Identifier`].
     RepositoryIdentifier(repository::identifier::Error),
 }
@@ -72,10 +65,6 @@ impl std::fmt::Display for Error {
             Error::Config(err) => write!(f, "configuration file error: {err}"),
             Error::Io(err) => write!(f, "i/o error: {err}"),
             Error::GrammarV1(err) => write!(f, "grammar error: {err}"),
-            Error::Repository(err) => write!(f, "repository error: {err}"),
-            Error::RepositoryBuilder(err) => {
-                write!(f, "repository builder error: {err}")
-            }
             Error::RepositoryIdentifier(err) => {
                 write!(f, "repository identifier error: {err}")
             }
@@ -94,11 +83,11 @@ type Result<T> = std::result::Result<T, Error>;
 #[command(author, version, about, long_about)]
 pub struct Args {
     /// The GitHub repositories to evaluate (e.g., "stjudecloud/workflows").
+    /// This will create temporary shallow clones of every
+    /// test repository specified on the CL. Normally, there is only one
+    /// repository on disk at a time. The difference in disk space usage
+    /// should be negligible.
     pub repositories: Option<Vec<String>>,
-
-    /// The location of the cache directory.
-    #[arg(long)]
-    pub cache_dir: Option<PathBuf>,
 
     /// The location of the config file.
     #[arg(short, long)]
@@ -130,9 +119,12 @@ pub struct Args {
     #[arg(short, long)]
     pub quiet: bool,
 
-    /// Overwrites the configuration file.
+    /// Overwrites the configuration file with new expected concerns and the
+    /// latest commit hashes. This will create temporary shallow clones of every
+    /// test repository. Normally, there is only one repository on disk at a
+    /// time. The difference in disk space usage should be negligible.
     #[arg(long)]
-    pub save_config: bool,
+    pub refresh: bool,
 
     /// Skips the retreiving of remote objects.
     #[arg(long)]
@@ -169,44 +161,40 @@ pub async fn gauntlet(args: Args) -> Result<()> {
         }
     };
 
+    let mut work_dir = WorkDir::default();
+
+    if args.refresh {
+        info!("refreshing repository commit hashes.");
+        config.inner_mut().update_repositories(work_dir.root());
+    }
+
     if let Some(repositories) = args.repositories {
-        config.inner_mut().extend_repositories(
-            repositories
-                .into_iter()
-                .map(|value| {
-                    value
-                        .parse::<Identifier>()
-                        .map_err(Error::RepositoryIdentifier)
-                })
-                .collect::<Result<Vec<_>>>()?,
-        )
+        repositories.into_iter().for_each(|value| {
+            let identifier = value
+                .parse::<Identifier>()
+                .map_err(Error::RepositoryIdentifier)
+                .unwrap();
+            work_dir.add_by_identifier(&identifier);
+        });
+        config
+            .inner_mut()
+            .extend_repositories(work_dir.repositories().clone())
     }
 
     let mut report = Report::from(std::io::stdout().lock());
 
-    for (index, repository_identifier) in config.inner().repositories().iter().enumerate() {
-        let mut repository =
-            repository::Builder::default().identifier(repository_identifier.clone());
-
-        if let Some(ref root) = args.cache_dir {
-            let mut repository_cache_root = root.clone();
-            repository_cache_root.push(repository_identifier.organization());
-            repository_cache_root.push(repository_identifier.name());
-            repository = repository.root(repository_cache_root);
-        }
-
-        if args.skip_remote {
-            let options = options::Builder::default().hydrate_remote(false).build();
-            repository = repository.options(options)
-        }
-
-        let mut repository = repository.try_build().map_err(Error::RepositoryBuilder)?;
-        let results = repository.hydrate().await.map_err(Error::Repository)?;
+    for (index, (repository_identifier, repo)) in config.inner().repositories().iter().enumerate() {
+        let results = repo.wdl_files(work_dir.root());
 
         report.title(repository_identifier).map_err(Error::Io)?;
         report.next_section().map_err(Error::Io)?;
 
-        for (path, content) in results {
+        for (relative_path, content) in results {
+            let abs_path = work_dir
+                .root()
+                .join(repository_identifier.organization())
+                .join(repository_identifier.name())
+                .join(&relative_path);
             if let Some(ref filters) = args.filter {
                 let mut skip = true;
 
@@ -218,15 +206,15 @@ pub async fn gauntlet(args: Args) -> Result<()> {
                 }
 
                 if skip {
-                    trace!("skipping: {path}");
+                    trace!("skipping: {:?}", abs_path);
                     continue;
                 }
             }
 
-            trace!("processing: {path}");
+            trace!("processing: {:?}", abs_path);
 
             let document_identifier =
-                document::Identifier::new(repository_identifier.clone(), path);
+                document::Identifier::new(repository_identifier.clone(), relative_path);
 
             match config.inner().version() {
                 wdl_core::Version::V1 => {
@@ -236,24 +224,63 @@ pub async fn gauntlet(args: Args) -> Result<()> {
                         .map_err(Error::GrammarV1)?
                         .into_parts();
                     if let Some(these_concerns) = these_concerns {
-                        detected_concerns.extend(these_concerns.into_inner().map(|concern| {
-                            ReportableConcern::from_concern(
-                                document_identifier.to_string(),
-                                concern,
-                            )
-                        }));
+                        if let Some(parse_errors) = these_concerns.parse_errors() {
+                            detected_concerns.extend(parse_errors.into_iter().map(|error| {
+                                // SAFETY: We ensure these concerns are not LintWarnings,
+                                // so they will always unwrap.
+                                ReportableConcern::from_concern(
+                                    document_identifier.to_string(),
+                                    wdl_core::Concern::ParseError(error.to_owned()),
+                                )
+                                .unwrap()
+                            }));
+                        }
+                        if let Some(validation_failures) = these_concerns.validation_failures() {
+                            detected_concerns.extend(validation_failures.into_iter().map(
+                                |failure| {
+                                    // SAFETY: We ensure these concerns are not LintWarnings,
+                                    // so they will always unwrap.
+                                    ReportableConcern::from_concern(
+                                        document_identifier.to_string(),
+                                        wdl_core::Concern::ValidationFailure(failure.to_owned()),
+                                    )
+                                    .unwrap()
+                                },
+                            ));
+                        }
                     }
 
                     if let Some(pt) = pt {
                         let (_, these_concerns) =
                             ast::v1::parse(pt).map_err(Error::AstV1)?.into_parts();
                         if let Some(these_concerns) = these_concerns {
-                            detected_concerns.extend(these_concerns.into_inner().map(|concern| {
-                                ReportableConcern::from_concern(
-                                    document_identifier.to_string(),
-                                    concern,
-                                )
-                            }));
+                            if let Some(parse_errors) = these_concerns.parse_errors() {
+                                detected_concerns.extend(parse_errors.into_iter().map(|error| {
+                                    // SAFETY: We ensure these concerns are not LintWarnings,
+                                    // so they will always unwrap.
+                                    ReportableConcern::from_concern(
+                                        document_identifier.to_string(),
+                                        wdl_core::Concern::ParseError(error.to_owned()),
+                                    )
+                                    .unwrap()
+                                }));
+                            }
+                            if let Some(validation_failures) = these_concerns.validation_failures()
+                            {
+                                detected_concerns.extend(validation_failures.into_iter().map(
+                                    |failure| {
+                                        // SAFETY: We ensure these concerns are not LintWarnings,
+                                        // so they will always unwrap.
+                                        ReportableConcern::from_concern(
+                                            document_identifier.to_string(),
+                                            wdl_core::Concern::ValidationFailure(
+                                                failure.to_owned(),
+                                            ),
+                                        )
+                                        .unwrap()
+                                    },
+                                ));
+                            }
                         }
                     }
 
@@ -344,7 +371,7 @@ pub async fn gauntlet(args: Args) -> Result<()> {
         .flatten()
         .collect::<IndexSet<_>>();
 
-    if args.save_config {
+    if args.refresh {
         info!("adding {} new expected concerns.", unexpected.len());
         info!(
             "removing {} outdated expected concerns.",
@@ -357,14 +384,14 @@ pub async fn gauntlet(args: Args) -> Result<()> {
             .chain(unexpected.iter())
             .cloned()
             .collect::<IndexSet<_>>();
-
         config.inner_mut().set_concerns(new);
+
         config.save().map_err(Error::Config)?;
     } else if !missing_but_expected.is_empty() {
         println!(
             "\n{}\n",
             "undetected but expected concerns remain: you should remove these from your \
-             configuration file or run the command with the `--save-config` option!"
+             configuration file or run the command with the `--refresh` option!"
                 .red()
                 .bold()
         );